+++ examples/development/simplecms_development_example.py
@@ -1,4 +1,10 @@
-# coding: utf-8
+#!/usr/bin/env python
+#coding: utf-8
+"""
+Simple Python CMS Instanz  - Test- und Entwicklung
+
+Simple Python CMS Instanz konfigurieren und Server starten
+"""
 
 import os
 import sys
@@ -8,20 +14,21 @@ THISDIR = os.path.dirname(os.path.abspath(__file__))
 try:
     import simplecms
 except ImportError:
-    # SIMPLECMSDIR und `sys.path.insert` werden nur zum Testen benötigt.
+    # `sys.path.insert` wird nur zum Testen benötigt.
     # Wenn `simplecms` installiert wurde, kann man es einfach importieren,
     # ohne vorher den Pfad angeben zu müssen.
-    SIMPLECMSDIR = os.path.abspath(os.path.join(THISDIR, "..", ".."))
-    sys.path.insert(0, SIMPLECMSDIR)
+    sys.path.insert(0, os.path.abspath(os.path.join(THISDIR, "..", "..")))
     import simplecms
 
-DATADIR = os.path.join(THISDIR, "data")
+
+# Konfiguriert die "Simple Python CMS"-Instanz und startet den Webserver
+simplecms.SimpleCms(
+    host = "127.0.0.1",
+    port = 8080,
+    datadir = os.path.join(THISDIR, "data"),
+).start()
 
 
-def main():
 
-    pass
 
-if __name__ == "__main__":
-    main()